Allow develop per spec instead of per package
Environments can contain the same package multiple times, but `develop`
only allows for one package to be developed (= built from local sources)
at a time.

This commit allows one develop entry per spec, and also makes the
develop yaml syntax equivalent to externals in packages. The new syntax
is:

```
develop:
- spec: mypkg@develop-a
- spec: mypkg@develop-b
  path: some/path
```

With this change the following use cases are covered:

- Updating a dependency of a package where you want to keep both the
stable version and the develop version around, and easily check if
tests of the parent still pass both for stable and dev version of the
dependency from a single environment
- Building different variants of the same package from the same local
sources (so you can make your changes in 1 place and test multiple
versions of it provided you can do out of source builds of course)
- Easily comparing two implementations of something, where you dev the
same package twice in different folders and you mark them with
`@develop-feature-a` and `@develop-feature-b`.

For concretization and determining the path of the sources we simply
take the first matching spec from the ordered `develop` array.